cmd/dba-qa/requirements.txt
click           >= 6.0
mysql-connector >= 2.1
pyyaml          >= 3.0
